project-list command and check correct session

+++ server/src/main/java/ru/trandefil/tm/service/ProjectServiceImpl.java
@@ -5,6 +5,7 @@ import ru.trandefil.tm.api.ProjectRepository;
 import ru.trandefil.tm.api.ProjectService;
 import ru.trandefil.tm.entity.Project;
 import ru.trandefil.tm.entity.Session;
+import ru.trandefil.tm.util.SignatureUtil;
 import ru.trandefil.tm.util.UUIDUtil;
 
 import java.util.List;
@@ -29,23 +30,39 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     public List<Project> getAll(Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            return null;
+        }
         return projectRepository.getAll();
+
     }
 
     public Project getById(String name, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            return null;
+        }
         return projectRepository.getById(name);
     }
 
     public void delete(Project project, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            return;
+        }
         projectRepository.delete(project);
     }
 
     public void deleteByName(String projectName, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            return;
+        }
         projectRepository.deleteByName(projectName);
     }
 
     @Override
     public Project getByName(String projectName, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            return null;
+        }
         return projectRepository.getByName(projectName);
     }
